Added to read me

+++ src/main/java/frc/robot/commands/MoveLift.java
@@ -59,7 +59,7 @@ public class MoveLift extends Command { //Moves the lift to the desired position
   @Override
   protected void execute() {
     int threshold = 1000;
-    int liftSpeed = 0.1;
+    double liftSpeed = 0.1;
     long current = Robot.m_lift.getEncoderMotor().getSelectedSensorPosition(); //Get current position
     long error = ticks - current; //Find the error of the current position and our desired position
